# backend/docs/swagger.yaml
components:
  schemas:
...
    chapters.CreateChapterRequest:
      properties:
        content:
          type: string
        display_order:
          type: integer
        status:
          type: string
        title:
          type: string
        volume_id:
          type: integer
        work_id:
          type: integer
      required:
      - title
      - work_id
      type: object
...
    chapters.UpdateChapterRequest:
      properties:
        content:
          type: string
        display_order:
          type: integer
        status:
          type: string
        title:
          type: string
        volume_id:
          type: integer
      type: object
...
    data:
      properties:
        data:
          properties:
            message:
              type: string
          type: object
      type: object
...
    response.StandardResponse:
      allOf:
      - $ref: '#/components/schemas/data'
      properties:
        code:
          type: integer
        data: {}
        message:
          type: string
        sourceId:
          type: string
        traceId:
          type: string
      type: object
...
  securitySchemes:
    BearerAuth:
      in: header
      name: Authorization
      type: apiKey
...
info:
  description: This is the API for Test.
  title: API 1.0
  version: "1.0"
openapi: 3.1.0
paths:
...
  /chapters:
    get:
      description: Get a list of all chapters for a specific work with pagination
      parameters:
      - description: Work ID
        in: query
        name: work_id
        required: true
        schema:
          type: integer
      - description: Page number
        in: query
        name: page
        schema:
          default: 1
          type: integer
      - description: Number of items per page
        in: query
        name: limit
        schema:
          default: 10
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                properties:
                  code:
                    type: integer
                  data: {}
                  message:
                    type: string
                  sourceId:
                    type: string
                  traceId:
                    type: string
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: work_id is required or invalid
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Failed to retrieve chapters
      security:
      - BearerAuth: []
      summary: List all chapters for a work
      tags:
      - chapters
    post:
      description: Create a new chapter for a work
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/chapters.CreateChapterRequest'
        description: Create Chapter Request
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                properties:
                  code:
                    type: integer
                  data: {}
                  message:
                    type: string
                  sourceId:
                    type: string
                  traceId:
                    type: string
                type: object
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Invalid request body
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Failed to create chapter
      security:
      - BearerAuth: []
      summary: Create a new chapter
      tags:
      - chapters
  /chapters/{id}:
    delete:
      description: Delete a chapter by its ID
      parameters:
      - description: Chapter ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                properties:
                  code:
                    type: integer
                  data: {}
                  message:
                    type: string
                  sourceId:
                    type: string
                  traceId:
                    type: string
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Invalid ID
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Chapter not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Failed to delete chapter
      security:
      - BearerAuth: []
      summary: Delete a chapter
      tags:
      - chapters
    get:
      description: Get a single chapter by its ID
      parameters:
      - description: Chapter ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                properties:
                  code:
                    type: integer
                  data: {}
                  message:
                    type: string
                  sourceId:
                    type: string
                  traceId:
                    type: string
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Invalid ID
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Chapter not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Failed to get chapter
      security:
      - BearerAuth: []
      summary: Get a single chapter
      tags:
      - chapters
    put:
      description: Update a chapter with the given details
      parameters:
      - description: Chapter ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/chapters.UpdateChapterRequest'
        description: Update Chapter Request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                properties:
                  code:
                    type: integer
                  data: {}
                  message:
                    type: string
                  sourceId:
                    type: string
                  traceId:
                    type: string
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Invalid ID or request body
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Chapter not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Failed to update chapter
      security:
      - BearerAuth: []
      summary: Update a chapter
      tags:
      - chapters
...